Avoid flagging setTimeout(..., 0) in lint
There isn't any difference between setTimeout(foo, 0) and step_timeout(foo, 0),
so it doesn't make sense to fail the lint in either case. Unfortunately getting
this perfectly correct is a bit tricky since there can be arbitary js in the
first argument to setTimeout. So doing it correctly requires a full js parser,
but we can do something with an approach that simply counts parens, skipping
strings, to find the end of the arguments.

This won't handle all cases, particularly involving template literals, but
may be better than nothing. In particular expressions like

setTimeout(() => `${"`"}`, 0}

will break since we won't skip the embedded expression in the template
literal.